--- a/src/core/ActionAtomistic.h
+++ b/src/core/ActionAtomistic.h
@@ -271,6 +271,22 @@ void ActionAtomistic::setGlobalPosition(const std::pair<std::size_t, std::size_t
   zpos[a.first]->data[a.second]=pos[2];
 }
 
+inline
+Vector ActionAtomistic::getForce( const std::pair<std::size_t, std::size_t>& a ) const {
+  Vector f;
+  f[0]=xpos[a.first]->getForce(a.second);
+  f[1]=ypos[a.first]->getForce(a.second);
+  f[2]=zpos[a.first]->getForce(a.second);
+  return f;
+}
+
+inline
+void ActionAtomistic::addForce( const std::pair<std::size_t, std::size_t>& a, const Vector& f ) {
+  xpos[a.first]->addForce( a.second, f[0] );
+  ypos[a.first]->addForce( a.second, f[1] );
+  zpos[a.first]->addForce( a.second, f[2] );
+}
+
 
 }
 #endif