@gtribello I tried to optimize wholemolecules, it's mostly inlining stuff but you might want to check if I did things right.

What worries me a bit is that even with this impressive improvement (almost a factor 2) I am back to timings pre-htt. Here's my timing (no graphs, but it should be easy to understand):
```
pre htt: 00b63d594487f4445e7d6748ef65c246c7ce3173

PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     2.426496     2.426496     2.426496     2.426496
PLUMED: 1 Prepare dependencies                          2000     0.001444     0.000001     0.000000     0.000012
PLUMED: 2 Sharing data                                  2000     0.493404     0.000247     0.000217     0.001881
PLUMED: 3 Waiting for data                              2000     0.000590     0.000000     0.000000     0.000013
PLUMED: 4 Calculating (forward loop)                    2000     1.376356     0.000688     0.000623     0.002047
PLUMED: 5 Applying (backward loop)                      2000     0.452501     0.000226     0.000200     0.002366
PLUMED: 6 Update                                        2000     0.002034     0.000001     0.000000     0.000078

current master 05fb5480bd42a1d077f564c55460c0d0961469fd

PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     4.537703     4.537703     4.537703     4.537703
PLUMED: 1 Prepare dependencies                          2000     0.002650     0.000001     0.000001     0.000027
PLUMED: 2 Sharing data                                  2000     0.651106     0.000326     0.000256     0.003563
PLUMED: 3 Waiting for data                              2000     0.002185     0.000001     0.000000     0.000048
PLUMED: 4 Calculating (forward loop)                    2000     2.870725     0.001435     0.001229     0.009485
PLUMED: 5 Applying (backward loop)                      2000     0.890085     0.000445     0.000349     0.004437
PLUMED: 6 Update                                        2000     0.002864     0.000001     0.000001     0.000035

optimize-wholemolecules 2daf02676cba141417062cea84baec51590f8b70 

PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     2.434463     2.434463     2.434463     2.434463
PLUMED: 1 Prepare dependencies                          2000     0.001600     0.000001     0.000000     0.000065
PLUMED: 2 Sharing data                                  2000     0.494312     0.000247     0.000216     0.000901
PLUMED: 3 Waiting for data                              2000     0.000606     0.000000     0.000000     0.000008
PLUMED: 4 Calculating (forward loop)                    2000     1.387576     0.000694     0.000602     0.001377
PLUMED: 5 Applying (backward loop)                      2000     0.450181     0.000225     0.000197     0.002094
PLUMED: 6 Update                                        2000     0.002222     0.000001     0.000001     0.000013
```

This is my input file:
```
WHOLEMOLECULES ENTITY0=1-100000
pos: POSITION ATOM=1
RESTRAINT ARG=pos.x AT=0.0 KAPPA=1
```

So far so good, but previous @gtribello 's tests were reporting wholemolecules as being [faster in htt than in master](https://github.com/plumed/plumed2/pull/933#issuecomment-1879661884), whereas I see it slower by a factor ~ 2x (before this modifications).

Is there something that I am missing? I am doing my tests on an intel MacBook Pro.