@gtribello 2024-02-15
@GiovanniBussi I had at this and I think it is fine to merge it.  It looks like it is the same.

I don't know why my computer was not reporting the slow down for WHOLEMOLECULES.  Perhaps my compiler automatically inlined those functions?