@GiovanniBussi 2024-02-15
Maybe it can inline across objects. Did you try on the ARM Mac? I will test with mine as well.