@gtribello 2024-02-15
I tested it on the ARM Mac.  it is the only one I have :-(