@GiovanniBussi 2024-02-19
I actually managed to try. Very interesting:
- in the test Gareth was doing (without applying force) I also see htt to be faster than pre-htt, largely because the `update` phase don't do anything in htt (I guess this optimization was missing before). Calculate is instead slightly slower in htt. This is consistent with what Gareth has seen before.
- in the test I was doing (applying force) I see htt to be slower than pre-htt
- my optimization merged in this PR are pretty useless on M2 processor

This is very weird. It looks like:
- the missing inlining per se is little relevant (calculate itself is not affected by inlining)
- application of forces is still a bit suboptimal in htt (there might be more work to do).
- putting this together with the timings I did on an intel mac, it looks like optimizations might have very different effects on different machines.

Raw data below.
```
bash-3.2$ cat pre
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     1.551619     1.551619     1.551619     1.551619
PLUMED: 1 Prepare dependencies                          2000     0.000081     0.000000     0.000000     0.000009
PLUMED: 2 Sharing data                                  2000     0.172640     0.000086     0.000085     0.000242
PLUMED: 3 Waiting for data                              2000     0.000051     0.000000     0.000000     0.000000
PLUMED: 4 Calculating (forward loop)                    2000     0.992900     0.000496     0.000489     0.000566
PLUMED: 5 Applying (backward loop)                      2000     0.314155     0.000157     0.000154     0.000210
PLUMED: 6 Update                                        2000     0.000054     0.000000     0.000000     0.000007
bash-3.2$ cat master
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     1.327825     1.327825     1.327825     1.327825
PLUMED: 1 Prepare dependencies                          2000     0.000158     0.000000     0.000000     0.000001
PLUMED: 2 Sharing data                                  2000     0.185173     0.000093     0.000092     0.000193
PLUMED: 3 Waiting for data                              2000     0.000146     0.000000     0.000000     0.000009
PLUMED: 4 Calculating (forward loop)                    2000     1.096683     0.000548     0.000543     0.000621
PLUMED: 5 Applying (backward loop)                      2000     0.000077     0.000000     0.000000     0.000001
PLUMED: 6 Update                                        2000     0.000061     0.000000     0.000000     0.000003
bash-3.2$ cat opt
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     1.328747     1.328747     1.328747     1.328747
PLUMED: 1 Prepare dependencies                          2000     0.000151     0.000000     0.000000     0.000001
PLUMED: 2 Sharing data                                  2000     0.187392     0.000094     0.000093     0.000196
PLUMED: 3 Waiting for data                              2000     0.000129     0.000000     0.000000     0.000001
PLUMED: 4 Calculating (forward loop)                    2000     1.095795     0.000548     0.000542     0.000596
PLUMED: 5 Applying (backward loop)                      2000     0.000096     0.000000     0.000000     0.000000
PLUMED: 6 Update                                        2000     0.000054     0.000000     0.000000     0.000002
bash-3.2$ cat pre.force
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     1.519698     1.519698     1.519698     1.519698
PLUMED: 1 Prepare dependencies                          2000     0.000092     0.000000     0.000000     0.000000
PLUMED: 2 Sharing data                                  2000     0.171945     0.000086     0.000085     0.000245
PLUMED: 3 Waiting for data                              2000     0.000050     0.000000     0.000000     0.000001
PLUMED: 4 Calculating (forward loop)                    2000     0.990990     0.000495     0.000492     0.000564
PLUMED: 5 Applying (backward loop)                      2000     0.311235     0.000156     0.000153     0.000229
PLUMED: 6 Update                                        2000     0.000073     0.000000     0.000000     0.000005
bash-3.2$ cat master.force 
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     1.789899     1.789899     1.789899     1.789899
PLUMED: 1 Prepare dependencies                          2000     0.000223     0.000000     0.000000     0.000001
PLUMED: 2 Sharing data                                  2000     0.187675     0.000094     0.000093     0.000193
PLUMED: 3 Waiting for data                              2000     0.000144     0.000000     0.000000     0.000002
PLUMED: 4 Calculating (forward loop)                    2000     1.149530     0.000575     0.000562     0.000631
PLUMED: 5 Applying (backward loop)                      2000     0.407120     0.000204     0.000202     0.000246
PLUMED: 6 Update                                        2000     0.000109     0.000000     0.000000     0.000006
bash-3.2$ cat opt.force 
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     1.764365     1.764365     1.764365     1.764365
PLUMED: 1 Prepare dependencies                          2000     0.000219     0.000000     0.000000     0.000001
PLUMED: 2 Sharing data                                  2000     0.185834     0.000093     0.000092     0.000196
PLUMED: 3 Waiting for data                              2000     0.000135     0.000000     0.000000     0.000002
PLUMED: 4 Calculating (forward loop)                    2000     1.128362     0.000564     0.000555     0.000620
PLUMED: 5 Applying (backward loop)                      2000     0.404132     0.000202     0.000200     0.000243
PLUMED: 6 Update                                        2000     0.000113     0.000000     0.000000     0.000006
```