@GiovanniBussi 2024-02-20
I think you are confused by my mistake above, I made a mess sorry. 

I ran them again now, I get something like this on INTEL. Timings on ARM are similar (same "ranking", but the differences between branches are someway smaller).

**without htt stuff**

```
no force, pre htt:

PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     2.769249     2.769249     2.769249     2.769249
PLUMED: 1 Prepare dependencies                          2000     0.000914     0.000000     0.000000     0.000021
PLUMED: 2 Sharing data                                  2000     0.541431     0.000271     0.000201     0.000891
PLUMED: 3 Waiting for data                              2000     0.000609     0.000000     0.000000     0.000023
PLUMED: 4 Calculating (forward loop)                    2000     1.577261     0.000789     0.000632     0.001633
PLUMED: 5 Applying (backward loop)                      2000     0.526821     0.000263     0.000202     0.000754
PLUMED: 6 Update                                        2000     0.000821     0.000000     0.000000     0.000074

force, pre htt:
PLUMED:                                                    1     2.777388     2.777388     2.777388     2.777388
PLUMED: 1 Prepare dependencies                          2000     0.001080     0.000001     0.000000     0.000022
PLUMED: 2 Sharing data                                  2000     0.524391     0.000262     0.000199     0.000941
PLUMED: 3 Waiting for data                              2000     0.000480     0.000000     0.000000     0.000021
PLUMED: 4 Calculating (forward loop)                    2000     1.613729     0.000807     0.000661     0.002137
PLUMED: 5 Applying (backward loop)                      2000     0.521073     0.000261     0.000201     0.000836
PLUMED: 6 Update                                        2000     0.000708     0.000000     0.000000     0.000041
```

Here you can see that pre-htt there's no significant optimization in place where you do not add any force. Also the cost of calculate is not affected, as expected.

**with htt stuff**

```
no force master (post htt)
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     3.054768     3.054768     3.054768     3.054768
PLUMED: 1 Prepare dependencies                          2000     0.001544     0.000001     0.000000     0.000023
PLUMED: 2 Sharing data                                  2000     0.559304     0.000280     0.000205     0.000992
PLUMED: 3 Waiting for data                              2000     0.001210     0.000001     0.000000     0.000024
PLUMED: 4 Calculating (forward loop)                    2000     2.377854     0.001189     0.000981     0.002303
PLUMED: 5 Applying (backward loop)                      2000     0.000898     0.000000     0.000000     0.000018
PLUMED: 6 Update                                        2000     0.000822     0.000000     0.000000     0.000284

force master (post htt)
PLUMED:                                               Cycles        Total      Average      Minimum      Maximum
PLUMED:                                                    1     4.343856     4.343856     4.343856     4.343856
PLUMED: 1 Prepare dependencies                          2000     0.002183     0.000001     0.000000     0.000081
PLUMED: 2 Sharing data                                  2000     0.656710     0.000328     0.000248     0.004149
PLUMED: 3 Waiting for data                              2000     0.001657     0.000001     0.000000     0.000035
PLUMED: 4 Calculating (forward loop)                    2000     2.612328     0.001306     0.001071     0.005927
PLUMED: 5 Applying (backward loop)                      2000     0.942530     0.000471     0.000371     0.004881
PLUMED: 6 Update                                        2000     0.001132     0.000001     0.000000     0.000021
```

The optimization is instead present in master (post htt merge). Notice  that also "calculate" slows down when you apply forces, I am not sure why (it didn't happen in pre htt). This optimization makes the overall performance without force closer to pre htt. Still, calculate is slower in post htt than in pre htt.

When using ARM, the slow down of calculate that I was seeing was smaller (~ 10%) and was cancelled by the optimization. I think this is why you reported htt to be faster than pre-htt with WHOLEMOLECULES (without forces).

I hope this is clearer (and there are no other mistakes).

In any case, the results (with forces) are significant on ARM as well, which means you can try yourself. What you should do is just:
- install somewhere a reference branch. I am using this as "pre htt": 00b63d594
- get current `master`, compile it, and source it
- run something like:
```
plumed-runtime benchmark --kernel /path/to/lib/libplumedKernel.dylib
```
with the kernel you want to test. If you use the installed master (and not the "sourceme") you have to do:
```
export PLUMED_KERNEL=/path/to/master/lib/libplumedKernel.dylib
/path/to/master/lib/plumed/plumed-runtime benchmark --kernel /path/to/lib/libplumedKernel.dylib
```

You need these tricks to run a different version of plumed within master